--- app/_components/AnimationWrapper/index.tsx
@@ -1,7 +1,0 @@
-type Props = {
-  children: React.ReactNode
-}
-
-export default function AnimationWrapper({ children }: Props) {
-  return <div className="animate-fade-in-bottom">{children}</div>
-}
