+++ app/components/ConvertDate/index.tsx
@@ -1,0 +1,16 @@
+import dayjs from 'dayjs'
+
+type Props = {
+  convertDate: string | number | Date
+  className?: string
+}
+
+export default function ConvertDate({ convertDate, ...props }: Props) {
+  const publishedAt = dayjs(convertDate).format('YYYY/MM/DD')
+
+  return (
+    <time dateTime={convertDate.toString()} {...props}>
+      {publishedAt}
+    </time>
+  )
+}
