+++ app/constants/index.ts
@@ -1,1 +1,14 @@
 export const WORKS_LIST_LIMIT = 6
+
+export const BIO = [
+  { id: 1, date: '2017.04', body: '高校卒業後、公務員として従事。' },
+  {
+    id: 2,
+    body: 'ホームページの管理業務を機にWebサイト制作に興味を持ち始め、独学で学習を始める。',
+  },
+  {
+    id: 3,
+    date: '2022.04〜',
+    body: 'ECコンサル会社のコーダーとして転職し、Web制作業務に携わる。',
+  },
+]
