+++ app/globals.css
@@ -4,8 +4,10 @@
 
 @layer base {
   :root {
+    --header-height: 72px;
+
     --background: 0 0% 100%;
-    --foreground: 222.2 84% 4.9%;
+    --foreground: 240 15.29% 15.88%;
 
     --card: 0 0% 100%;
     --card-foreground: 222.2 84% 4.9%;
@@ -13,16 +15,16 @@
     --popover: 0 0% 100%;
     --popover-foreground: 222.2 84% 4.9%;
 
-    --primary: 221.2 83.2% 53.3%;
+    --primary: 173.41 80.39% 40%;
     --primary-foreground: 210 40% 98%;
 
     --secondary: 210 40% 96.1%;
     --secondary-foreground: 222.2 47.4% 11.2%;
 
     --muted: 210 40% 96.1%;
-    --muted-foreground: 215.4 16.3% 46.9%;
+    --muted-foreground: 240 5.2% 33.92%;
 
-    --accent: 210 40% 96.1%;
+    --accent: 207 30% 92.1%;
     --accent-foreground: 222.2 47.4% 11.2%;
 
     --destructive: 0 84.2% 60.2%;
@@ -36,7 +38,7 @@
   }
 
   .dark {
-    --background: 0 0% 20%;
+    --background: 240 6% 10%;
     --foreground: 0 0% 98%;
 
     --card: 222.2 84% 4.9%;
@@ -45,22 +47,22 @@
     --popover: 222.2 84% 4.9%;
     --popover-foreground: 210 40% 98%;
 
-    --primary: 217.2 91.2% 59.8%;
+    --primary: 173.41 80.39% 40%;
     --primary-foreground: 222.2 47.4% 11.2%;
 
     --secondary: 217.2 32.6% 17.5%;
     --secondary-foreground: 210 40% 98%;
 
-    --muted: 217.2 17.18% 13.58%;
-    --muted-foreground: 0 0% 74%;
+    --muted: 231.43 9.86% 13.92%;
+    --muted-foreground: 240 33.33% 94.12% / 70%;
 
-    --accent: 217.2 32.6% 17.5%;
+    --accent: 221.2 11.6% 17.5%;
     --accent-foreground: 210 40% 98%;
 
     --destructive: 0 62.8% 30.6%;
     --destructive-foreground: 210 40% 98%;
 
-    --border: 0 0% 13.18%;
+    --border: 0 0% 18%;
     --input: 217.2 32.6% 17.5%;
     --ring: 212.7 26.8% 83.9%;
   }
@@ -71,7 +73,10 @@
     @apply border-border;
   }
   body {
-    @apply bg-background text-foreground;
+    @apply bg-background text-muted-foreground font-base;
+  }
+  h1,h2,h3 {
+    @apply text-foreground;
   }
   h1 {
     @apply text-4xl font-semibold tracking-wide md:text-[40px] pb-6;
@@ -82,7 +87,13 @@
   h3 {
     @apply font-medium pb-4 tracking-wide;
   }
+  p {
+    @apply text-muted-foreground;
+  }
   section {
     @apply pb-16;
+  }
+  h2:where(section > *) {
+    @apply pt-2 pb-8;
   }
 }
