--- app/utils/motion.ts
@@ -1,16 +1,0 @@
-export const fadeUp = {
-  hidden: {
-    opacity: 0,
-    y: '25px',
-    transition: {
-      duration: 0.7,
-    },
-  },
-  visible: {
-    opacity: 1,
-    y: 0,
-    transition: {
-      duration: 0.7,
-    },
-  },
-}
