+++ app/work/layout.tsx
@@ -1,9 +1,7 @@
-import Container from '../_components/Container'
-
 type Props = {
   children: React.ReactNode
 }
 
-export default function RootLayout({ children }: Props) {
-  return <Container>{children}</Container>
+export default function WorkLayout({ children }: Props) {
+  return <>{children}</>
 }
