+++ components.json
@@ -11,7 +11,7 @@
     "prefix": ""
   },
   "aliases": {
-    "components": "@/app/_components",
+    "components": "@/app/components",
     "utils": "@/lib/utils"
   }
 }
