+++ package.json
@@ -20,37 +20,38 @@
     "@types/react": "18.2.20",
     "@types/react-dom": "18.2.7",
     "autoprefixer": "10.4.15",
-    "cheerio": "^1.0.0-rc.12",
+    "cheerio": "1.0.0-rc.12",
     "class-variance-authority": "^0.7.0",
-    "clsx": "^2.1.0",
+    "clsx": "^2.1.1",
     "date-fns": "^3.6.0",
-    "dayjs": "^1.11.9",
-    "highlight.js": "^11.8.0",
+    "dayjs": "^1.11.11",
+    "framer-motion": "^11.2.4",
+    "highlight.js": "^11.9.0",
     "hljs": "^6.2.3",
     "husky": "^8.0.3",
     "install": "^0.13.0",
-    "lint-staged": "^15.1.0",
+    "lint-staged": "^15.2.2",
     "lucide-react": "^0.354.0",
-    "microcms-js-sdk": "^2.6.1",
-    "next": "^13.5.2",
+    "microcms-js-sdk": "^2.7.0",
+    "next": "^14.2.3",
     "next-themes": "^0.2.1",
     "optional": "^0.1.4",
     "plaiceholder": "^3.0.0",
-    "pnpm": "^8.12.1",
+    "pnpm": "^8.15.8",
     "postcss": "8.4.28",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-icons": "^4.10.1",
-    "react-intersection-observer": "^9.5.2",
-    "sharp": "^0.33.1",
-    "tailwind-merge": "^2.2.1",
+    "react-icons": "^4.12.0",
+    "react-intersection-observer": "^9.10.2",
+    "sharp": "^0.33.3",
+    "tailwind-merge": "^2.3.0",
     "tailwindcss": "3.3.3",
     "tailwindcss-animate": "^1.0.7",
     "typescript": "5.1.6"
   },
   "devDependencies": {
     "@biomejs/biome": "1.5.3",
-    "@cloudflare/next-on-pages": "^1.6.0"
+    "@cloudflare/next-on-pages": "^1.11.3"
   },
   "packageManager": "pnpm@8.12.1",
   "lint-staged": {
